lib: add dump files for v20.0 ABI
This patch adds the lib .dump files generated for the v20.0 ABI. These
files will be used to compare the ABI during the meson build when the
compat_checks meson is enabled. If there is a mismatch between the .dump
files (the baseline, ie. current stable ABI version) and the code being
compiled, then abidiff will report an error and the build will fail.

The .dump files are stored in the lib/.abi directory.  The directory is
hidden to reduce clutter and also to discourage changing these files since
they should not be modified.

Signed-off-by: Kevin Laatz <kevin.laatz@intel.com>